rename : BookStatus -> BookStatusType
