+++ Library/src/main/java/com/programmers/library/domain/Book.java
@@ -1,14 +1,14 @@
 package com.programmers.library.domain;
 
 
-import static com.programmers.library.domain.BookStatus.RENTABLE;
+import static com.programmers.library.domain.BookStatusType.RENTABLE;
 
 public class Book {
     private final Long bookId;
     private final String title;
     private final String author;
     private final Integer page;
-    private BookStatus bookStatus;
+    private BookStatusType bookStatus;
 
     public Book(Long bookId, String title, String author, Integer page) {
         this.bookId = bookId;
@@ -17,7 +17,7 @@ public class Book {
         this.page = page;
         this.bookStatus = RENTABLE;
     }
-    public Book(Long bookId, String title, String author, Integer page, BookStatus bookStatus){
+    public Book(Long bookId, String title, String author, Integer page, BookStatusType bookStatus){
         this.bookId = bookId;
         this.title = title;
         this.author = author;
@@ -25,7 +25,7 @@ public class Book {
         this.bookStatus = bookStatus;
     }
 
-    public void updateBookStatus(BookStatus bookStatus){
+    public void updateBookStatus(BookStatusType bookStatus){
         this.bookStatus = bookStatus;
     }
     public Long getBookId() {
@@ -40,7 +40,7 @@ public class Book {
     public Integer getPage() {
         return page;
     }
-    public BookStatus getBookStatus() {
+    public BookStatusType getBookStatus() {
         return bookStatus;
     }
 }
