+++ Library/src/main/java/com/programmers/library/repository/Repository.java
@@ -1,7 +1,7 @@
 package com.programmers.library.repository;
 
 import com.programmers.library.domain.Book;
-import com.programmers.library.domain.BookStatus;
+import com.programmers.library.domain.BookStatusType;
 
 import java.util.List;
 import java.util.Optional;
@@ -13,7 +13,7 @@ public interface Repository {
     public void deleteBook(Long id);
     public List<Book> findBooksByTitle(String title);
     public Optional<Book> findBookById(Long id);
-    public void updateStatus(Book book, BookStatus bookStatus);
+    public void updateStatus(Book book, BookStatusType bookStatus);
     public Long findLastId();
 
 }
